kalshi-arb/pyproject.toml
[project]
name = "kalshi-arb"
version = "0.1.0"
description = "Structural arbitrage engine for Kalshi prediction markets"
requires-python = ">=3.11"
dependencies = [
    "pykalshi>=1.0.4",
    "websockets>=12.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "structlog>=24.1",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "tenacity>=8.2",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "rich>=13.7",
    "typer>=0.12",
    "msgpack>=1.0",
    "jinja2>=3.1",
    "itsdangerous>=2.1",
    "python-multipart>=0.0.9",  # required for fastapi form parsing on login
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.5",
    "types-pyyaml",
]
# libsql-experimental is the Turso driver. Only needed for the live-
# migration architecture (see docs/turso-setup.md). Paper phase uses
# local SQLite via the same StoreBackend abstraction, which is why
# this dependency is OPTIONAL rather than in the main list.
#
# The package has no pre-built wheel for some newer Python/platform
# combos -- installing it falls back to a from-source Rust build that
# requires cargo + a C toolchain on the box. Keep it out of the paper-
# phase critical path: the operator doesn't need Rust to run the bot.
#
# Install with:  pip install -e ".[turso]"
turso = [
    "libsql-experimental>=0.0.41",
]

[project.scripts]
kalshi-arb = "kalshi_arb.cli:app"

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["kalshi_arb*"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --strict-markers -m 'not live'"
markers = [
    "live: touches the public network (cloudflared quick-tunnel end-to-end). Opt-in.",
]
